code_taxonomy: Add feature to print tallies over a time period
I want to be able to make a "burndown" chart of code in a codebase, so I
can see how much work has been done in a period (and how much is left to
do).

This commit adds a new command line mode that will run the same search
multiple times in the same git repo(s), checking out different commits
for different points in the time period specified.

The cli interface is crap right now and will be changed, but it works,
so commiting it before I break it.